Lock Flutter version to 3.7.10
After reading the README.md and going over the project, I did not see
a specific Flutter version, so in this PR I set it to the latest stable
release, which is 3.7.10 as of today.

Other files updated were a product of running pub get after setting
the Flutter version.

<!-- ps-id: f84e4135-08df-439c-b2ff-4288b299b4f9 -->